moved api test functions to a separate file so I don't have to think about it
